ui 업데이트 및 프래그먼트 책임 분리
