+++ app/src/main/res/layout/fragment_growthalbum.xml
@@ -20,9 +20,9 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:background="@color/white"
+            android:onClick="@{() -> viewmodel.pickDate()}"
             android:translationZ="10dp"
             android:visibility="gone"
-            android:onClick="@{() -> viewmodel.pickDate()}"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
@@ -65,23 +65,11 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="5dp"
+            android:onClick="@{() -> viewmodel.pickDate()}"
             android:src="@drawable/baseline_keyboard_arrow_down_24"
             app:layout_constraintBottom_toBottomOf="@+id/tvDate"
             app:layout_constraintStart_toEndOf="@+id/tvDate"
             app:layout_constraintTop_toTopOf="@+id/tvDate" />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/appCompatButton3"
-            android:layout_width="wrap_content"
-            android:layout_height="40dp"
-            android:layout_marginEnd="40dp"
-            android:layout_marginBottom="10dp"
-            android:background="@drawable/button_show_date"
-            android:backgroundTint="@color/grey"
-            android:onClick="@{() -> viewmodel.pickDate()}"
-            android:text="@string/view_by_date"
-            app:layout_constraintBottom_toTopOf="@+id/myCalendar"
-            app:layout_constraintEnd_toEndOf="parent" />
 
 
         <com.michalsvec.singlerowcalendar.calendar.SingleRowCalendar
@@ -162,29 +150,6 @@
 
         </com.google.android.material.card.MaterialCardView>
 
-        <ImageView
-            android:id="@+id/floating_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="50dp"
-            android:backgroundTint="@color/baba_main"
-            android:src="@drawable/baby_button"
-            app:backgroundTint="@color/white"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cardView" />
-
-        <ImageView
-            android:id="@+id/imageView9"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="10dp"
-            android:src="@drawable/baby_button_icon"
-            app:backgroundTint="@color/white"
-            app:layout_constraintBottom_toBottomOf="@id/floating_button"
-            app:layout_constraintEnd_toEndOf="@id/floating_button"
-            app:layout_constraintStart_toStartOf="@+id/floating_button"
-            app:layout_constraintTop_toTopOf="@id/floating_button" />
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/baby_select_view"
@@ -193,8 +158,8 @@
             android:background="@drawable/white_card_view"
             android:clickable="true"
             android:elevation="10dp"
-            android:visibility="gone"
             android:focusable="true"
+            android:visibility="gone"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent">
